requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic-settings>=2.2,<3.0

pdfplumber>=0.11,<1.0

numpy>=1.26,<3.0
sentence-transformers>=3.0,<4.0
faiss-cpu>=1.8,<2.0

httpx>=0.27,<1.0